+++ src/automoveis.cpp
@@ -1,7 +1,7 @@
 #include "automoveis.h"
 using namespace std;
 
-Automoveis::Automoveis(string t_m, string marca, double preco, string chassi, int fabricacao) //construtor do novo carro
+Automoveis::Automoveis(string t_m, string marca, int preco, string chassi, int fabricacao) //construtor do novo carro
 {
     setTipo_motor(t_m);
     setMarca(marca);
